Make sure that element_id is a string in fetch_osm_element
It may not be obvious that fetch_osm_element wants element_id
to be a string, so convert it to a string anyway.  If you
pass an integer then the element you're looking for won't
be found in the dictonaries built by the parser, where the
keys are strings, and the server will get hit again and
again looking for it.