Add a callback option to OSMXMLParser for lower memory use
The old mode of operation of the parser was to parse every
element in the file and then return them.  This option allows
you to instead supply a callback which will be called with each
top-level element as it's parsed.  If the parser is used with
such a callback, the list of top level elements is no longer
kept, since typically you would use the callback instead for
parsing very large files, and you're trying to use minimal
memory.