Consider a point in each polygon when finding a parent
When importing boundary line data from scratch using
bdline_gb-2012-05.zip, the parent for the civil parish Magor with
Undy Community isn't found because the point picked by
point_on_surface from the first polygon in the shapefile isn't
contained in Monmouthshire Council's boundary. This commit considers
all polygons in the child shapefile until a point which is contained
in a parent boundary is found. Presumably this is a mistake in the
boundary data, but this seems like a reasonable workaround.