Compare newly generated KML files with the current generation in MapIt Global
This script, when given a directory that contains a newly generated
set of KML files, will check every KML file against the database to
see whether the boundary has changed, just appeared, is invalid, etc.
and produce a CSV file with a row for each KML file.

This is useful for working out how many of the boundaries have
changed since the last import, and so how much of an increase
in database size one might expect on importing it.