Add a script for timing point -> areas lookups in MapIt UK
This script generates random-enough locations in the UK
and uses the timeit module to time Area.objects.by_location
lookups for those locations.  This is useful to assess how
the performance of MapIt UK would be affected by adding
polygons of various sizes and complexities to an existing
instance.

This script was written with Jenny Duckett.