Fix the MapIt Global import script for importing a second generation
This script had previously only been tested with the initial data
import of MapIt Global, and it was undecided at that stage how to
deal with the next import (e.g. whether to preserve the boundaries
from the previous generation or not).

Our decision was to preserve the previous generation; we test
whether the new boundary is the same as that in the current
generation in the database - if so, then the generation_high
for that area is extended (and the names are updated) or
otherwise a new generation is created.

The changes to the import script in this commit puts this decision
into effect.