Fix an error with an initial Boundary-Line import without --commit
Thanks to Geoff Stockham for pointing out this problem.  This
fixes issue #48.

When we see a second polygon with the same unit_id and this
is a new import, a dry-run import will fail because the name
can't be found in the database from the unit_id - it was never
saved the first time around.  As with finding the name from
ons_code we should ignore this error if we're not saving to the
database.

(The example in the bug report is the second polygon in
Waterbeach ED in county_electoral_division_region.shp.)