Move KML, GeoJSON and WKT generation into a method of Area
It is useful to be able to export an area in these different
formats in situations other than a view, e.g. from a django-admin
command.  This refactoring moves the code that generates KML,
GeoJSON and WKT into a method of Area, which doesn't depend on
having a request object.

This refactoring also makes it easier to detect when your
simplify_tolerance is high enough that all the polygons disappear,
in which case a SimplifiedAway exception is raised.