Add a command for generating CSV for Google Fusion Tables
It's often useful to be able to easily visualize a particular
subset of areas from MapIt on a map.  This command
(mapit_make_fusion_csv) lets you specify a particular area
type and generate a CSV file that can be imported into
Google Fusion Tables to show all areas of the that type
randomly coloured.  You can also restrict this to only show
areas of that type that are covered by another area.

Some instructions for importing the CSV into Google Fusion
Tables is given in the comments at the top of the file
and I will create a similar page on code.mapit.mysociety.org