Add an option to create a new area if polygons have changed
The default behaviour of the generic import script, if it finds
an area with the same name / code in a previous generation, is
to replace the polygons associated with that area.

However, we may want alternative behaviour, where if a new
generation of data has a changed boundary for an area whose
name / code remains the same, we create a new area with those
polygons and preserve the previous one.  This commit introduces
that change with the --preserve option.

Fixes #52