mapit_import: show a percentage representing change in the boundaries
When --verbose is specified, this will also show the percentage
difference between (a) the area of the symmetric difference
between the old boundary and the new boundary and (b) the area
of the old boundary.  This is a reasonable indication of how
much an area with the same name / code has changed from one
generation to the next.