Allow much longer names for areas
There are cases in both MapIt Global and police neighbourhood data of
names that are much longer than the current limit of 100 characters.
Since PostgreSQL only uses the minimum storage necessary for varchars
regardless of their maximum length, there's no harm in making the
max_length for the name field of Name much larger.  This commit raises
it to 2000.

(An alternative would be to just switch the name field to be a
TextField rather than a CharField, but I'm keeping this as a CharField
so that the field's default appearance in the admin interface remains
unchanged.)