Add options to mapit_import to allow overriding of name and code
Names and codes for areas are not always present as fields within
data files and they may need to be imported from other sources,
for example filenames or data fetched via APIs. The options
added in this commit allow a name and a code to be specified
to use for areas to be imported from a file. A warning is shown
if multiple areas are found in the file, which would create
areas with duplicate names and/or codes.