Add a script to import police force area boundaries
Police force boundaries for England, Wales and Northern Ireland
are now published as KML files under the UK Open Government
Licence at:

  http://data.gov.uk/dataset/police-force-boundaries-england-and-wales

This dataset does not include names or codes for the areas
within the KML files. The filenames correspond to the force
IDs used by the police API, so they are stored as codes and
the force names are fetched and saved via the API at:

  http://data.police.uk/api/forces

This script calls the generic mapit_import script to handle
the KML files and save the areas. It accepts and passes on
all of that command's behaviour-related options.