Add an option to mapit_import for a field to suffix to the name
In some datasets you need to compose two fields to produce a unique
name.  For example, in the case of the South African Wards
shapefile there are no unique ward names - the best one can do is
to make a ward name from the municipality name suffixed by the ward
number.

It's debatable whether this is the right approach, or whether,
for example, one should allow the --name_field option to take
multiple fields separated by a separator character (',' or '|',
perhaps) but composing two fields is probably enough for most cases
and hazarding a character that won't appear in field names in
ESRI shapefiles or KML files isn't trivial.