Restore optional use of a remote Overpass server
When we switched to using a local Overpass server for MapIt
Global, the option to use a remote Overpass server instead
was removed.  This restores that option, controlled by settings
in conf/general.yml (LOCAL_OVERPASS, OVERPASS_DB_DIRECTORY
and OVERPASS_SERVER.)  It's still useful to be able to use
the remote server out-of-the-box for testing, and quickly
grabbing the KML for and OSM way or relation.

Thanks to Cesar Martinez Izquierdo (@dispiste) who made a
pull request with similar functionality some time ago, but
who closed it (unmerged) later.  This commit introduces similar
changes, but using urllib2 instead of relying on wget, and
using the MapIt general.yml configuration file instead of a
new configuration module.