Add a setting / context variable for whether this MapIt uses postcodes
We're shortly going to add another country that we don't currently have
postcode data for, and currently the test for whether that's the case
is just "country == 'Global'" in the templates - that can't be easily
extended because of the limitations on condidations in Django templates
and it would also be bad to have to repeat that logic everywhere.

Instead, this commit adds a settings variable POSTCODES_AVAILABLE and
it's also added to the country context processor so that it's available
in templates as postcodes_available.