Include properties with geojson, ensure FeatureCollection
This pulls in an area's properties into the geojson
so that they don't also need to call .json for each area.

This also ensures that /areas/1.geojson always returns a
FeatureCollection, which means the caller/client isn't surprised that
sometimes the response is a polygon, and sometimes it's a feature
collection.  They always knows what type of object is going to come back
and can act accordingly. If they know they only want a single item, they
can use /area/1.geojson.