Add --only-with to mapit_import so you can only import particular areas
Sometimes one has to deal with a shapefile which has many features in
it, but you only want to import some subset based on key / value
combinations of their fields.  This commit adds the --only-with option
to let you specify that only features with particular properties should
be imported from the file. For example, you could do:

    ./manage.py mapit_import \
        --only-with=TIPO:CIUDAD \
        --only-with=PROVINCIA:'BUENOS AIRES' \
        ...
        010_Ejidos_Urbanos.shp

... to only import the boundaries of cities in the province of Buenos
Aires.