Add a Procfile
We need this to deploy MapIt as a normal app in our infrastructure.

This file is very similar to the one for Stagecraft:

    https://github.com/alphagov/stagecraft/commit/3eff938ad04427758e66599bac0b6c6811e4d41c

MapIt gets less than 5 requests a second on average, and each request
completes in well under half a second, so 8 workers on each machine is
generous. We need a single machine to be able to handle all production traffic
while we update the database using our current process, so this seems
reasonable.