Replace mySociety README with a GOV.UK specific one
For GOV.UK apps the README is usually a quick reference for developers who
may not be familiar with the project.

Add a new README that follows the GOV.UK README styleguide [1], aiming to
orient those new to MapIt and its use in GOV.UK.

The original (`.rst`) README was removed as GH still picks that over the
new (`.md`) if both exist. To avoid confusion the new links to the old
on the mySociety repo.

[1]: https://github.com/alphagov/styleguides/