Add a command to raise the generation_high on current areas
This is useful if you've created a new generation for a new data
import, but there are existing areas in the database that should
still remain active after you've activated your new generation.
After creating the new inactive generation you can run this command
to raise the generation_high of any areas who have a generation_high
set to the current (i.e. most recent active) generation so that
they're the newly created generation.

You could do this in the Django shell, but this seems like a safer
way of doing this.

This includes options for country / type restriction, as suggested
by Matthew Somerville