Add script for adding missing codes to areas
To get the data, first we ran mapit_UK_show_missing_codes against the
rebuilt database and compared with running the same script against the
production data from S3 (based on the dump in puppet[1]).

We then looked at various sources to work out what these differences
are and if we can sort them.  Mostly this was looking at the history of
the s_mapit_server in puppet [2] (of particular interest this commit to
add the gss code for Northern Ireland[3] and this commit to re-add some
missing ons codes[4]).

We also looked at the notes on the mapit-scripts[5] and used @daivaughan's
local authorities site[6] to fill in the blanks of the few remaining.

For the ons codes that are missing it's worth reading the Mapit notes on
ids[7] to get some clarity on why they might be missing.  ONS stopped
producing these codes in 2012, so when boundaries changed for some areas
they get a new gss code, but not a new ons code.  We've decided (as
mapit have) to find the ons code from the old equivalent area and apply
it to the new area.  This *may* not be exactly correct, but as we still
use these codes, it's a pragmatic approach until we no longer require them.

[1]: https://github.gds/gds/puppet/blob/6fbd021e0f9344526f61f2b8556aee1697a8023f/modules/govuk/manifests/node/s_mapit_server.pp#L24
[2]: https://github.gds/gds/puppet/commits/master/modules/govuk/manifests/node/s_mapit_server.pp
[3]: https://github.gds/gds/puppet/commit/6fbd021e0f9344526f61f2b8556aee1697a8023f
[4]: https://github.gds/gds/puppet/commit/66ff9b9506ef93d6274ded22c263148288cb1400
[5]: https://github.gds/gds/mapit-scripts/blob/master/README.md#notes
[6]: http://govuklocal.dafyddvaughan.co.uk/authorities/
[7]: https://github.com/mysociety/mapit/wiki/NotesOnIDs