Script for importing UK postcodes from ONSPD
Importing everything from one source reduces what we have to download
and we have to download ONSPD to get NI and Crown dependencies which
aren't in Code Point Open.

Other advantages are that the ONSPD has all live and terminated
postcodes in it and (as of Aug 2015 release at least) everything is a
gss code, rather than a mix of ons + gss.

We add a toggle to allow importing terminated postcodes as it can be
useful to have old postcodes in the db to allow searching on old
addresses.  Note that although code point open doesn't include
terminated postcodes we can end up with them in our dataset, but only if
we had a long-lived database that imported multiple releases of the
dataset.  For example, we import the May 2012 dataset and then when the
next one comes out in Aug 2012 we import that - our db will have in it
the postcodes that were terminated between those two releases, but it
won't have any that were terminated before May 2012.  This leads to the
situation where a db rebuilt from scratch using the current dataset
would have a different set of postcodes to one that had been around for
a few years having had releases imported as they arrive.  Notionally
both represent the current data, but one has more postcodes.  Using the
ONSPD and allowing terminated postcodes fixes this problem.