Provide fixture data for NI council / electoral / wards
In April 2015 councils and wards changed in Northern Ireland so the old ni-electoral-areas data files no longer represent the truth.

The new ni-electoral-areas-2015 file provides the names and GSS codes of the new Districts, Electoral Areas, and Wards of Northern Ireland following the Apr 2015 reorganisation.  We synthesized this from a few datasets.
** The District -> Electoral Area -> Ward breakdown is taken directly from the legislation[2] - although this only contains the names
** The GSS codes of the Districts and Wards are taken from the "Wards (2015) to district council areas (2015) NI lookup"[3] dataset provided by the ONS.
** The GSS codes of the Electoral areas are taken from running sparql queries against the ONS Linked Data Portal[4].  The query used was:

    PREFIX rdf: <http://www.w3.org/1999/02/22-rdf-syntax-ns#>
    PREFIX geography: <http://statistics.data.gov.uk/def/>
    PREFIX statistical-entity: <http://statistics.data.gov.uk/def/statistical-entity#>
    SELECT DISTINCT ?item

    WHERE {
    ?item rdf:type geography:statistical-geography .

      ?item statistical-entity:code <http://statistics.data.gov.uk/id/statistical-entity/N10> .
    }
    ORDER BY ASC(?code)
    LIMIT 100
    OFFSET 0

N10 is the code for a NI Electoral Area (found by looking up one of the areas by name and investigating).  We then extract the data for each entry in this result set and extract the GSS code to match up with the name.

This data will be used to help import NI areas from the OSNI boundary datasets which do not all contain GSS codes.

[1]: https://geoportal.statistics.gov.uk/geoportal/catalog/search/resource/details.page?uuid=%7B2196C1D5-6A11-47DE-BD0E-26311E3D6D9F%7D
[2]: http://www.legislation.gov.uk/uksi/2014/270/made
[3]: https://geoportal.statistics.gov.uk/geoportal/catalog/search/resource/details.page?uuid=%7BFE83C43C-9403-408C-833C-367BC56659C9%7D
[4]: http://statistics.data.gov.uk/