Allow importing postcodes with no location
Some rows in the ONSPD that have no location - those where the 12th
column is a '9'.  The existing importers for ni and gb postcodes
automatically ignore these, but it can be useful to include them for
existence checks even if we can't give more geographical information
about them.

We implement `location_available_for_row` to return `False` if the
quality row is `'9'`, `True` otherwise.  This lets us import rows that
have no location if we want to.  To keep the Code-Point Open import
behaviour we provide an option `--allow-no-location-postcodes` to turn
this on; without this option the importer will not import these rows.