Add detecting location availability per postcode.
If `--no-location` is not set we would try to detect a location for each
row, and this would break if the location fields could not be coerced
into floats.  Some datasets mix location and non-location postal codes
and to import them all we have to filter the data and run the importer
twice.

This change allows individual importers to implement
`location_available_for_row` to say if the supplied row has location
data or not.  The method is called on each row and will run the
`--no-location` path if we can't extract location fields for that row.
If `--no-location` is set, we always run that path, regardless of the
`location_available_for_row` value.