[UK] Handle NI postcodes in ONSPD importer
We add an option to allow importing the NI postcodes at the same time as
the rest of the postcodes.  The option takes 3 values: 'include',
'exclude', 'only' with the same behaviour as the --crown-dependencies
option:

* 'include' will import NI postcodes
* 'exclude' will not import NI postcodes
* 'only' will only import NI postcodes

The default is 'exclude' to maintain previous behaviour.  Unlike Crown
Dependency postcodes NI postcodes might have location data, and so the
--allow-no-location-postcodes setting (default false) does affect how we
import NI postcodes.

Setting both --crown-dependencies and --northern-ireland to 'only' is an
error and will halt the importer before it begins.

We've also updated the documentation provided by the options to be
clearer about how the various options interact.

Unlike the old nspd_ni importer that relied on the nspd_ni_areas
importer to be run and the ni-electoral-areas.csv to directly assign
areas to NI postcodes, this importer has no special handling.  We assume
that the new OSNI importer has been run and the relevant shapefiles have
been imported, much like we assume that the boundary-line importer has
been run to provide the areas for the rest of the UK.

You can also specify srids for GB/NI postcodes separately (with sensible
defaults), to change the srid option on a per-row basis.