Add script for importing a new db from the S3 dump
On the old mapit servers we did this via a combination of a fabric script and puppet.  Puppet downloaded the dump and loaded it into the DB, fabric wiped the old db and then ran puppet.  We'll still need a fabric script to orchestrate runnin this script on servers (and to bounce the app / nginx to pick up the new db and data).  We moved to an app-local script because:

1. loading data via puppet is not something that other govuk-puppet-managed-apps do.
2. it breaks the requirement on deploying puppet if we want to load a new data set into deployed environments.  Note that instead we have a requirement on deploying the app itself because the url and sha are part of the script; they could be arguments but we wanted a git-based audit trail of when the dataset is updated.
3. it gives us a loading script that will work on the dev vm allowing us to load the data for development.

Note that the dataset and sha referenced are made up of the ONSPD August 2015 dataset, Boundary Line November 2015 dataset, and the NI shape data from November 2015.  See: https://github.gds/gds/mapit-scripts/pull/2, https://github.com/alphagov/mapit/pull/7 and https://trello.com/c/a7c1LPbl/26-get-mapit-database-onto-new-preview-machine for more context on this import process and data.