[UK] Add command for importing OSNI boundary data.
This is mostly a rip from import_boundary_line but instead of taking one
file as an argument we specify which files to work on as options because
the OSNI releases are single files per shape, unlike Boundary-Line which
has everything in one folder.  We could expect a user to put everything in
one folder first, but we need to handle each shapefile differently
depending on how we expect it to be used.  For example the Westminster
Parliamentary constituencies boundaries have the name in PC_NAME and the
gss code in PC_ID, whereas the LGD file has them in LGDNAME and LGDCode.

We add a new codetype for identifying the OSNI id of the boundaries and a
new nametype for the OSNI names.

Notes:
* The ward boundaries data was published once in a set that had no
WardCode, but did have LGDName, so we could use our 2015 fixture to do a
name match to get the GSS code. Hopefully though they won't publish that
dataset again.
* The Euro boundary import generates some warnings about invalid
geometry, but the existing "fix_invalid_geos_geometry" is able to turn
the data into something that it considers valid.