Add option for importing OSNI ward boundaries
Name is in WARDNAME and GSS code is in WardCode.  We did see them publish
this data in a set that had no WardCode, but did have LGDName so we could
use our ni-electoral-areas-2015.csv fixture to do a name match to get the
GSS code.  Hopefully though they won't publish that dataset again.