Import NI Electoral Areas (LGE)
These are the areas that live between LGD (NI Councils) and LGW (NI
Wards).  The shapefile released by the OSNI does not include GSS codes
so we have to synthesize them by doing a name lookup against the
ni-electoral-areas-2015.csv fixture.  We've asked OSNI if they plan to
expose GSS codes in this dataset.