[UK] Allow specifying SRID of OSNI imports.
From looking at the shapefiles in a viewing tool (qgis) it appears that
the co-ords are in the NI projection (29902) in some files, but in the
102100 projection in others. For some files this matches up with the
arcgis metadata pointed to by the OSNI for each release (e.g. the LGEs
dataset[1] points to the following metadata as its source[2] which lists
29902 as the projection, and the LGW dataset[3] points to source
metadata[4] which lists 102100 projection). For others however this is
not true (e.g. the LGDs dataset[5] has a source metadata[6] that says
29902, but the downloaded shapefile is actually in 102100).

As it appears this is changeable per release (possibly per download) we
allow for telling the importer what srid a given file is in. We convert
everything to the MAPIT_AREA_SRID before import.

Note that for some reason the geometry imported from the shapefiles does
not contain an SRID for some reason so we have to set it even if we
don't have to transform it.

As a further wrinkle, PostGIS doesn't support 102100, but it is
mathematically equivalent to 3857 which it does support. Unfortunately
using that projection causes failures during for point-based lookup of
parents, but if we use 4326 instead it works. Apparently 102100 and 4326
are both "web mercator" projections so are probably very similar (if not
exactly mathematically equivalent). Interestingly opening a shapefile
that is in 102100 in a viewing tool such as qgis reports it as 4326
whereas a 29902 reports as a custom projection that is identical in all
but name to 29902. This suggests it's safe to use 4326 as a replacement
for 102100.

The defaults we set for the options are based on the SRIDs of the data
files we've downloaded in Dec 2015 - they may change over time.

[1]: http://osni.spatial-ni.opendata.arcgis.com/datasets/981a83027c0e4790891baadcfaa359a3_4
[2]: https://gisservices.spatialni.gov.uk/arcgisc/rest/services/OpenData/OSNIOpenData_LargescaleBoundaries/MapServer/4
[3]: http://osni.spatial-ni.opendata.arcgis.com/datasets/55cd419b2d2144de9565c9b8f73a226d_0
[4]: https://services3.arcgis.com/dNsInyVNGMqG1QjF/arcgis/rest/services/OSNI_Open_Data_Largescale_Boundaries_Wards_2012/FeatureServer/0
[5]: http://osni.spatial-ni.opendata.arcgis.com/datasets/a55726475f1b460c927d1816ffde6c72_2
[6]: https://gisservices.spatialni.gov.uk/arcgisc/rest/services/OpenData/OSNIOpenData_LargescaleBoundaries/MapServer/2