[UK] Add script for adding GSS codes to NI Areas
Some mapit installations already have NI Areas with or without boundaries
but these areas may not have GSS codes.  This script uses the
ni-electoral-areas-2015.csv hierarchy to find LGDs, LGEs, and LGWs by
name and add their GSS codes.  Because names are not neccessarily unique
it respects the hierarchy in the fixture.  If names cannot be found (it
does a case-insensitive lookup) the row is ignored and a warning issued.