Add script for adding names to NI Areas
The names in the OSNI data don't always match the names for the same
areas in the ni-electoral-areas-2015.csv fixture which was extracted from
the legislation.  In some cases it's just an uppercase difference, or a
lack of punctuation.  In others the names are completely different.  For
example in the fixture GSS N09000011 is called "North Down and Ards" but
in the OSNI shapefile it is called "East Coast".  Turns out this is
because the council voted to change the name to the OSNI one, but backed
down after outcry and reverted[1].

This script goes through the fixture and matches on GSS code to find the
Areas and add a new override name to the area if the fixture name is not
already present.

[1]: http://www.belfasttelegraph.co.uk/news/northern-ireland/backlash-forces-council-to-ditch-new-east-coast-name-that-cost-thousands-30902221.html