Correct LGD names in ni-electoral-areas-2015.csv
Mostly this is just extending the name to include the council type (District, Borough, or City), similar to naming of some council areas in the rest of the UK.  In the case of "North Down and Ards" we also rename to their final name choice of "Ards and North Down".  For "Derry and Strabane" and "Armagh, Banbridge and Craigavon" we also include "City" in the appropriate place ("Derry City" and "Armagh City").