Convert geometry to application projection in NI shape imports
We used to import all the NI shapes with an srid of 29902 (the Irish grid [1]).
For some reason when the geometry was extracted from the DB it was in 27700
(the GB grid [2]) but had not undergone any transformation from 29902 to 27700.
Consequently the NI shapes were in the wrong place (covering Liverpool, North
Wales and some of the Irish Sea).

It's not clear how this happened, but we can fix it by always transforming the
NI shape data from whatever srid it is provided as into the 27700 srid used by
the rest of the UK data.  Note that we actually use the
`settings.MAPIT_AREA_SRID` srid and not 27700 directly as in most cases of a UK
instance of mapit this will be 27700, but in the off chance it's not we don't
want things to break.

[1]: http://spatialreference.org/ref/epsg/tm65-irish-grid/
[2]: http://spatialreference.org/ref/epsg/27700/