Download the S3 db dump as the deploy user
The script is intended to be run as a user that can passwordless sudo so
they can become postgres and run the pg commands to create the db, install
the extensions, and import the db dump.  However, this user won't have
write permissions on the app directory so the download will fail.  Happily
they can also passwordless sudo to deploy who will have write permissions.

We don't need the sudo on reading the file as our non-sudo user will have
read permissions.