[UK] For NIE areas, allow import via name match.
If you're importing to an empty database, there's no problem, but you
might be importing into a database that already has NIE areas in it and
don't want to create new ones simply to add the new boundaries. These
are unable to match via ONS code, so instead allow a match on name for
NIE boundaries only.